Replace CLI smoke test with library integration test
Remove the CLI smoke test (cloud_cli) that provisioned real cloud
resources to test CLI wiring. The library integration test now covers
the full API lifecycle, and CLI-specific concerns (arg parsing, JSON
output, request building) are already covered by unit tests in
commands.rs.

- Delete crates/clickhousectl/tests/cloud_cli/ (test + support)
- Add library integration test and support module
- Update cloud-integration.yml to run library test, scope triggers
  to library crate only
- Add test-cli.yml for CLI build/clippy/unit tests on PRs
- Update README to reference the library integration test
