+++ .github/workflows/test-cli.yml
@@ -1,0 +1,31 @@
+name: Test CLI
+
+on:
+  workflow_dispatch:
+  pull_request:
+    paths:
+      - "crates/clickhousectl/**"
+      - "Cargo.lock"
+      - ".github/workflows/test-cli.yml"
+
+permissions:
+  contents: read
+
+jobs:
+  test:
+    name: Test clickhousectl
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install Rust toolchain
+        uses: dtolnay/rust-toolchain@stable
+
+      - name: Build
+        run: cargo build -p clickhousectl
+
+      - name: Clippy
+        run: cargo clippy -p clickhousectl -- -D warnings
+
+      - name: Test
+        run: cargo test -p clickhousectl
