+++ crates/clickhouse-cloud-api/tests/integration/mod.rs
@@ -1,0 +1,1 @@
+pub mod support;
