--- crates/clickhousectl/tests/cloud_cli.rs
@@ -1,5 +1,0 @@
-#[path = "cloud_cli/service_crud.rs"]
-mod service_crud;
-
-#[path = "cloud_cli/support.rs"]
-mod support;
